test(Cascara): [FDS-194] move things around for better cosmos experience
