+++ packages/cascara/src/ui/Table/fixtures/DeveloperExperience.fixture.js
@@ -1,48 +1,10 @@
-import React from 'react';
+import React, { useCallback } from 'react';
+import { Provider } from 'reakit';
+
 import Table from '../Table';
 import { results } from '../data/entities';
 
 import { ACTIONS, COLUMNS } from './constants';
-
-const Loading = () => (
-  <>
-    <h3>Loading</h3>
-    <p>
-      The table component will not blow up if there are no values defined. It
-      will also explicitly show a loading state if the data attribute is
-      undefined (or null).
-    </p>
-
-    <Table />
-  </>
-);
-
-const Empty = (fixtureProps) => (
-  <>
-    <h3>Empty</h3>
-    <p>
-      If a table gets an empty array, that should be treated different from
-      undefined or null. An empty array should mean that there are no records
-      and we should indicate this to users in the table itself.
-    </p>
-
-    <Table {...fixtureProps} />
-  </>
-);
-
-const DataOnly = (fixtureProps) => (
-  <>
-    <h3>Data Only</h3>
-    <p>
-      When only passing an array of data results to a Table, we should render
-      all results, even if no other props are defined yet. We should also
-      attempt to infer the correct module type to display. This includes tricky
-      types like object and array using a private module for JSON.
-    </p>
-
-    <Table {...fixtureProps} />
-  </>
-);
 
 const DataWithDisplay = (fixtureProps) => (
   <>
@@ -73,20 +35,79 @@ const DataWithDisplay = (fixtureProps) => (
       </ul>
     </div>
 
-    <Table {...fixtureProps} />
+    <Provider>
+      <Table {...fixtureProps} />
+    </Provider>
   </>
 );
 
-const UnknownModules = (fixtureProps) => (
+const UnknownActionModules = (fixtureProps) => (
+  <>
+    <h3>
+      actions.modules with <code>unknown modules</code>
+    </h3>
+    <p>
+      A helpful error message is displayed if an action module is not found.
+    </p>
+
+    <Provider>
+      <Table {...fixtureProps} />
+    </Provider>
+  </>
+);
+
+const UnknownDataModules = (fixtureProps) => (
   <>
     <h3>
       dataDisplay with <code>unknown modules</code>
     </h3>
-    <p>A helpful error message is displayed if a module is not found.</p>
+    <p>A helpful error message is displayed if a data module is not found.</p>
 
-    <Table {...fixtureProps} />
+    <Provider>
+      <Table {...fixtureProps} />
+    </Provider>
   </>
 );
+
+const WithDeprecatedProps = ({ onAction: extOnAction, ...props } = {}) => {
+  const onAction = useCallback((action, record) => {
+    console.log(`Action ${action.name} invoked`);
+
+    switch (action.name) {
+      case 'viewInFAQ':
+        console.table(record);
+        break;
+
+      case 'edit.start':
+        // do something
+        break;
+
+      case 'edit.cancel':
+        // do something
+        break;
+
+      case 'edit.save':
+        // do something
+        break;
+
+      default:
+        break;
+    }
+  }, []);
+
+  return (
+    <>
+      <h3>With deprecated props</h3>
+      <p>
+        Some props have been deprecated, this is a safeguard for developerss
+      </p>
+
+      <Provider>
+        <Table {...props} onAction={extOnAction || onAction} />
+      </Provider>
+    </>
+  );
+};
 
 // We export the data results here for use in our tests.
 // This also allows us to generate test data from inside a fixture and then reuse it in tests.
@@ -94,9 +115,6 @@ const dataResults = results;
 export { dataResults };
 
 export default {
-  loading: <Loading />,
-  empty: <Empty data={[]} uniqueIdAttribute='eid' />,
-  dataOnly: <DataOnly data={results} uniqueIdAttribute='eid' />,
   dataWithDisplay: (
     <DataWithDisplay
       data={results}
@@ -105,7 +123,7 @@ export default {
     />
   ),
   unknownActionModule: (
-    <UnknownModules
+    <UnknownActionModules
       actions={{
         ...ACTIONS,
         modules: [
@@ -120,10 +138,20 @@ export default {
     />
   ),
   unknownDataModule: (
-    <UnknownModules
+    <UnknownDataModules
       actions={ACTIONS}
       data={results}
       dataDisplay={[...COLUMNS, { attribute: 'eid', module: 'unknownModule' }]}
+      uniqueIdAttribute='eid'
+    />
+  ),
+  withDeprecatedProps: (
+    <WithDeprecatedProps
+      data={results}
+      dataConfig={{
+        actions: ACTIONS.modules,
+        display: COLUMNS,
+      }}
       uniqueIdAttribute='eid'
     />
   ),
